+++ pyasic/miners/device/models/auradine/AT/AT1.py
@@ -1,4 +1,4 @@
-from pyasic.device import MinerAlgo
+from pyasic.device.algorithm import MinerAlgo
 from pyasic.device.models import MinerModel
 from pyasic.miners.device.makes import AuradineMake
