+++ pyasic/miners/device/models/avalonminer/nano/nano3.py
@@ -1,4 +1,5 @@
-from pyasic.device import MinerAlgo, MinerModel
+from pyasic.device.algorithm import MinerAlgo
+from pyasic.device.models import MinerModel
 from pyasic.miners.device.makes import AvalonMinerMake
